apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
python-multipart
python-dotenv
chromadb
groq
google-generativeai
cohere
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
apscheduler
httpx
pypdf
pinecone
